requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langchain-groq
langchain-google-genai
langchain-huggingface
chromadb
langchain-chroma
sentence-transformers
pypdf
python-docx
python-pptx
ddgs
requests